Added SpringSecurity to both ocntroller

+++ src/main/java/com/ahmed/journalApp/service/JournalEntryService.java
@@ -6,6 +6,7 @@ import com.ahmed.journalApp.repo.JournalEntryRepo;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,17 +25,16 @@ public class JournalEntryService {
 
     @Transactional
     public void saveEntry(JournalEntry journalEntry,String username){
-        try{
             User user = userService.findByUsername(username);
+            if (user == null) {
+                throw new UsernameNotFoundException(
+                        "User not found: " + username
+                );
+            }
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepo.save(journalEntry);
             user.getJournalEntries().add(saved);
             userService.saveEntry(user);
-        } catch (Exception e) {
-           // log.error("Exception ",e);
-            System.out.println(e);
-            throw new RuntimeException("An error occuring while saving the entry",e);
-        }
     }
 
     public void saveEntry(JournalEntry journalEntry){
@@ -54,8 +54,47 @@ public class JournalEntryService {
 
     public  void deleteById(ObjectId id,String username){
         User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
+        if (user == null){
+            throw new RuntimeException("User not found");
+        }
+        boolean removed = user.getJournalEntries()
+                        .removeIf(entry-> entry.getId().equals(id));
+        if(!removed){
+            throw new RuntimeException("Journal does not belong to this user");
+        }
         userService.saveEntry(user);
         journalEntryRepo.deleteById(id);
     }
+
+    public JournalEntry updateJournalEntry(
+            ObjectId id,
+            JournalEntry newEntry,
+            String username) {
+
+        User user = userService.findByUsername(username);
+
+        if (user == null) {
+            return null;
+        }
+
+        boolean ownsJournal = user.getJournalEntries()
+                .stream()
+                .anyMatch(entry -> entry.getId().equals(id));
+
+        if (!ownsJournal) {
+            return null;
+        }
+
+        JournalEntry oldEntry = journalEntryRepo.findById(id).orElse(null);
+
+        if (oldEntry == null) {
+            return null;
+        }
+
+        oldEntry.setTitle(newEntry.getTitle());
+        oldEntry.setContent(newEntry.getContent());
+
+        return journalEntryRepo.save(oldEntry);
+    }
+
 }
